give up
On this sad day in history our hero was
defeated by the great moto.

The combination of moto mocking s3,
boto to talk to the mocked s3 and fsspec
being used for http tests makes the s3 tests
a rather large wrench to throw into the unit tests.

The issues are:

- s3 tests must be setup before http tests so that
  the mock s3 configuration is setup before fsspec
  is imported. This complicates otherwise simple
  http tests.
- moto boto and s3fs often result in conflicting
  dependencies (looking at you aiobotocore) which
  in the past created impossible environments and
  currently results in installing and uninstalling
  certain versions to find one that works.

Since we already have "compatibility_tests" I figured
we could make "integration_tests" which now include:

- compability tests (same as before)
- s3 tests (to run s3 tests in isolation for 1 environment)

If we assume fsspec is doing what it says on the tin
then testing http should be enough. I am in favor of
doing some sort of s3 test since that's the main target
fsspec backend (and not hadoop, arrow, etc). We're not
testing the other backends so I'm ok with just having
minimal s3 tests and relying on fsspec to test that it
supports the backends it claims.
