+++ .github/workflows/ci.yml
@@ -134,4 +134,4 @@ jobs:
       # Any env name which does not start with `pyXY` will use this Python version.
       default_python: '3.11'
       envs: |
-        - linux: py311-s3fs
+        - linux: s3
