+++ asdf/_tests/test_generic_io.py
@@ -261,7 +261,7 @@ def test_urlopen(tree, httpserver):
 
 
 @pytest.fixture(params=[True, False])
-def warn_no_fsspec(request, s3):
+def warn_no_fsspec(request):
     if request.param:
         yield nullcontext()
     else:
