+++ integration_tests/compatibility/assert_file_correct.py
