+++ tox.ini
@@ -1,8 +1,9 @@
 [tox]
 env_list =
     compatibility
+    s3
     coverage
-    py{39,310,311,312,313}{,-compatibility,-coverage,-jsonschema}{,-devdeps}{,-parallel}{,-pytestdev}{,-s3fs}
+    py{39,310,311,312,313}{,-compatibility,-coverage,-jsonschema}{,-devdeps}{,-parallel}{,-pytestdev}
     asdf{-standard,-transform-schemas,-unit-schemas,-wcs-schemas,-coordinates-schemas,-astropy,-zarr,-compression}
     astrocut
     gwcs
@@ -25,9 +26,10 @@ deps =
     oldestdeps: minimum_dependencies
     parallel: pytest-xdist
     pytestdev: git+https://github.com/pytest-dev/pytest
-    s3fs: moto[s3]
-    s3fs: s3fs
-    s3fs: flask-cors
+    s3: moto[s3,server]
+    s3: boto
+    s3: fsspec
+    s3: s3fs
 extras = all,tests
 # astropy will complain if the home directory is missing
 pass_env = HOME
@@ -47,12 +49,12 @@ commands_pre =
 commands =
     coverage: coverage run --source=asdf --rcfile={tox_root}/pyproject.toml -m \
     pytest \
-    compatibility: compatibility_tests/ \
+    compatibility: integration_tests/compatibility/ \
+    s3: integration_tests/s3/ \
     --remote-data \
     --durations=10 \
     jsonschema: --jsonschema \
     parallel: --numprocesses auto \
-    s3fs: asdf/_tests/test_s3fs.py
 # the OpenAstronomy workflow appends `--cov-report` in `{posargs}`, which `coverage` doesn't recognize
     !coverage: {posargs}
     coverage:
